Make existing render hotkeys configurable
